requirements.txt
pdfplumber>=0.11.0
wordninja>=2.0.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pandas>=2.2.0
streamlit>=1.33.0
openai>=1.0.0
google-genai>=1.0.0
